+++ sdks/universal-router-sdk/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @uniswap/universal-router-sdk
 
+## 5.8.0
+
+### Minor Changes
+
+- 4263dcf: Add Ink chain and Universal Router deployment config (v2.2.0 router also exposed as v2.1.1)
+
+### Patch Changes
+
+- Updated dependencies [4263dcf]
+  - @uniswap/sdk-core@7.18.0
+  - @uniswap/router-sdk@2.10.5
+  - @uniswap/v2-sdk@4.20.5
+  - @uniswap/v3-sdk@3.30.5
+  - @uniswap/v4-sdk@2.2.3
+
 ## 5.7.0
 
 ### Minor Changes