+++ sdks/v4-sdk/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @uniswap/v4-sdk
 
+## 2.2.3
+
+### Patch Changes
+
+- Updated dependencies [4263dcf]
+  - @uniswap/sdk-core@7.18.0
+  - @uniswap/v3-sdk@3.30.5
+
 ## 2.2.2
 
 ### Patch Changes